Fix up quan headers to get quan things working with 8 bit AVR e.g Arduino for the Arduino Pro Mini based GPS only tracker airborne Be explicit about int32 int64 etc remove some unnecessary cstdint includes Make quantity rational value_type int32_t for AVR else int64_t

+++ quan/tracker/zapp4/crc.hpp
@@ -1,7 +1,7 @@
 #ifndef QUAN_TRACKER_ZAPP4_CRC_HPP_INCLUDED
 #define QUAN_TRACKER_ZAPP4_CRC_HPP_INCLUDED
 
-#include <cstdint>
+#include <quan/config.hpp>
 
 namespace quan{ namespace tracker{ namespace zapp4{
 
@@ -10,11 +10,11 @@ namespace quan{ namespace tracker{ namespace zapp4{
    inline uint32_t crc32(uint32_t crc, uint32_t data)
    {
       crc = crc ^ data;
-      for ( uint32_t i = 0U; i < 32U; ++i){
-         if (crc & (1 << 31)){
-            crc =  (crc << 1) ^ 0x4c11db7;
+      for ( uint32_t i = 0U; i < static_cast<uint32_t>(32U); ++i){
+         if (crc & (static_cast<uint32_t>(1U) << static_cast<uint32_t>(31U))){
+            crc =  (crc << static_cast<uint32_t>(1U)) ^ static_cast<uint32_t>(0x4c11db7);
          }else{
-            crc <<= 1;
+            crc <<= static_cast<uint32_t>(1U);
          }
       }
       return crc;
